refactor(devtools): address review feedback on env/trustchain/bindings
- devtools/env: replace custom jest config with @support/jest-devtools preset
- devtools/trustchain: extract toHex/fromHex to utils.ts; introduce
  useEncryptUserData and useDecryptUserData shared hooks to deduplicate
  logic between .web and .native components
- devtools/bindings: derive TrustchainDevToolProps/CloudSyncDevToolProps
  via Extract<DevToolsConfig[number], {id}>['config'] from @devtools/registry,
  removing direct deps on @devtools/trustchain and @devtools/cloud-sync
