+++ devtools/trustchain/src/components/sdk/DecryptUserData.native.tsx
@@ -1,18 +1,8 @@
-import { useCallback, useState } from "react";
 import { TextInput } from "react-native";
 import { useTheme } from "@ledgerhq/lumen-ui-rnative";
 import type { Trustchain, TrustchainSDK } from "../../types";
 import { Actionable } from "../Actionable";
-
-const fromHex = (hex: string): Uint8Array => {
-  const clean = hex.replace(/\s/g, "");
-  if (clean.length % 2 !== 0 || !/^[0-9a-fA-F]*$/.test(clean)) {
-    throw new Error("Invalid hex string");
-  }
-  const bytes: number[] = [];
-  for (let i = 0; i < clean.length; i += 2) bytes.push(Number.parseInt(clean.slice(i, i + 2), 16));
-  return new Uint8Array(bytes);
-};
+import { useDecryptUserData } from "./useDecryptUserData";
 
 export function DecryptUserData({
   sdk,
@@ -22,23 +12,19 @@ export function DecryptUserData({
   trustchain: Trustchain | null;
 }>) {
   const { theme } = useTheme();
-  const [input, setInput] = useState("");
-  const [output, setOutput] = useState<string | null>(null);
-
-  const action = useCallback(
-    (tc: Trustchain, hex: string) =>
-      sdk.decryptUserData(tc, fromHex(hex)).then(arr => new TextDecoder().decode(arr)),
-    [sdk],
+  const { input, inputs, setInput, output, setOutput, action, valueDisplay } = useDecryptUserData(
+    sdk,
+    trustchain,
   );
 
   return (
     <Actionable
       buttonTitle="sdk.decryptUserData"
-      inputs={trustchain && input ? [trustchain, input] : null}
+      inputs={inputs}
       action={action}
       value={output}
       setValue={setOutput}
-      valueDisplay={v => v}
+      valueDisplay={valueDisplay}
     >
       <TextInput
         placeholder="hex to decrypt"
