+++ devtools/trustchain/src/components/sdk/EncryptUserData.native.tsx
@@ -1,13 +1,8 @@
-import { useCallback, useState } from "react";
 import { TextInput } from "react-native";
 import { useTheme } from "@ledgerhq/lumen-ui-rnative";
 import type { Trustchain, TrustchainSDK } from "../../types";
 import { Actionable } from "../Actionable";
-
-const toHex = (arr: Uint8Array): string =>
-  Array.from(arr)
-    .map(b => b.toString(16).padStart(2, "0"))
-    .join("");
+import { useEncryptUserData } from "./useEncryptUserData";
 
 export function EncryptUserData({
   sdk,
@@ -17,22 +12,19 @@ export function EncryptUserData({
   trustchain: Trustchain | null;
 }>) {
   const { theme } = useTheme();
-  const [input, setInput] = useState("");
-  const [output, setOutput] = useState<Uint8Array | null>(null);
-
-  const action = useCallback(
-    (tc: Trustchain, msg: string) => sdk.encryptUserData(tc, new TextEncoder().encode(msg)),
-    [sdk],
+  const { input, inputs, setInput, output, setOutput, action, valueDisplay } = useEncryptUserData(
+    sdk,
+    trustchain,
   );
 
   return (
     <Actionable
       buttonTitle="sdk.encryptUserData"
-      inputs={trustchain && input ? [trustchain, input] : null}
+      inputs={inputs}
       action={action}
       value={output}
       setValue={setOutput}
-      valueDisplay={v => toHex(v)}
+      valueDisplay={valueDisplay}
     >
       <TextInput
         placeholder="message to encrypt"
